FastSMS/requirements.txt
fastapi == 0.97.0
SQLAlchemy == 2.0.16
mysqlclient == 2.2.0
pydantic == 1.10.9
uvicorn == 0.22.0
passlib == 1.7.4
bcrypt == 4.0.1
python-jose == 3.3.0
python-multipart == 0.0.6
python-dateutil == 2.8.2